ml_project/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.78.0
debugpy==1.8.19
decorator==5.2.1
dnspython==2.8.0
docker==7.1.0
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==8.38.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.3.0
websockets==15.0.1
Werkzeug==3.1.5
zipp==3.23.0